backend/requirements.txt
Flask>=2.3.3
Flask-SQLAlchemy>=3.0.5
Flask-Migrate>=4.0.5
Flask-JWT-Extended>=4.5.3
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.4
Flask-Limiter>=3.5.0
python-dotenv>=1.0.0
marshmallow>=3.20.1
marshmallow-sqlalchemy>=0.29.0
bcrypt>=4.0.1
eventlet>=0.33.3
geopy>=2.3.0
psycopg2-binary>=2.9.7
python-dateutil>=2.8.2
email-validator>=2.0.0
sendgrid>=6.10.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
# Testing
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0